Notify players about next games and owner changes (refs #987, #985).
The table plugin modifications done previously for #985 have proved to be
buggy in subtle ways, so I had to rewrite them. I am much more satisfied with
the current approach than the previous one.

Instead of setting a new game as the current game, and later discarding it,
don't set a new game as the current game of the table until it is ready,
instead store all "next" games as pending games of the table and only "promote"
one of them to the current game when it reaches the 'invitation' state.

There are three types of dialogs that a player can see:

- A player has joined a pending game, and has now been chosen to be the next owner.
  He is prompted to create a new game.
- A player has joined a pending game, but another player created another playable game
  which is now the next game of the table. He is prompted to join it.
- Owner of a game in pending state is slow, and another player beats him at creating
  the next game. He is prompted to join the other game.

I was having problems with JS tests in IE - they were flip-flopping with random phantom
errors, which is something we have seen before. After struggling with that for a while,
I discovered the reason was some jqDock code that was running asynchronously, and kept
running past the end of associated test, messing with the state of other tests.

In order to get around this (to wait until jqDock has finished doing its thing
before going on to the next test), the $.cardstories.create function now returns
a deferred object in cases where we need them for testing.