Spin pending games into their own tables (closes #994).
When one of the pending games is promoted to become the next
game of the table, the remaining pending games are spun off
into their own tables.

I noticed there was a problem with a modal overlay blocking
access to the sentence box in create_write_sentence state,
so that was fix as well.