Add an option to turn animations off (refs #998).
If the parameter 'animations' with value 'off' is present in the url,
the animations are "skipped", allowing one to play the game much faster
which can be useful when doing manual tests.

This is achieved by turning jQuery animations off by setting $.fx.off = true
in combination with modifying calls to $.cardstories.delay and
$.cardstories.animationTimeout to always invoke the underlying function with
delay set to zero.

The function $.cardstories.animationTimeout was added. When animations are on,
it works exactly like $.cardstories.setTimeout, but when animations are off,
it always calls setTimeout with a zero delay. A new function was needed because
we use $.cardstories.setTimeout for tasks other than animations, where we can't
just change the delay to 0 and expect things to still work.

Another minor change I slipped into this commit is changing the timeout
after a failed ajax call retries. It was previously set to 100 ms. It was me
who initially chose that value, but to be honest, I have no idea why I thought
100 ms would be a good value. I bumped it up to 1 second so that when something
bad happens, the server isn't showered with neverending retry requests.
One second might still be to short, but I didn't want to change things drastically yet.