Fix a number of problems related to tables (refs #998).
The on_tab_closed hook was removed - we aren't using monitoring
of tab open/close events to detect player's presence anymore.
This was done for simplicity and because we are using timeouts
to cycle among potential next owner now, I believe this to be
less of an issue.

The other important change was delaying the time when pending games
are detached from the current table. Previously they were detached too
soon, now they are only detached if they reach 'invitation' state, or
if the current game goes into 'complete' state.

More info can be found in the comments on the ticket:

http://tickets.farsides.com/issues/998