Fix further problems with tables and next_owner_ids (closes #998).
After doing more thorough manual testing, I discovered two other problems:

1) Players who already joined one of the pending games wouldn't be given
   a chance to be the next owner.

The reason was the the update_next_owner_id() method would check the player's
opened tabs to see if player keeps the current game opened in a tab.
The reason was that when a player joins the pending game, the tab with the
previous game is automatically closed (and replaced with the pending game's tab).

To fix this problem I only removed the check for an opened tab. I believe that since
we now have timeouts, this shouldn't be too much of a problem anymore.
It would be possible to make the table plugin smarter by also checking for opened
pending games, but I decided to try the simplest thing first, and improve on it
in the future if it turns out to be neccessary.

While players who joined a pending games should be given a chance to be the
next owner, the owners of pending games shouldn't, so I had to add a simple filter
to filter out pending game owners when selecting suitable next owners.

2) When a player who joined a pending game was notified that he was the next
   owner, the text that says "<player name> took to long to create the next game",
   had the current player's name in the <player name> placeholder, while it should
   have been the name of the pending game owner that the current player has joined.