Remove the -e flag before django-paypal package in requirements.txt.
It's supposed to be used when you want to hack on the package, as it
checks out the code into a 'src' folder under the current dir.
We don't really want that, just install django-paypal into the site-packages dir.

For some reason, without the -e flag, I had to specify the protocol
as git+https, otherwise I would get errors when pip was trying to split
the string at '+'. Not sure why it worked with the -e flag, though.