Implement a proof-of-concept page with a paypal 'Buy Now' button. (refs #1001).
I implemented a simple page with a 'Buy Now' button.
Upon clicking the button, the user is taken to the Paypal page, where he can
make the purchase. After the purchase is made, the user is redirected back to our
website, with PDT variables in the URL.

The django-paypal module takes care of retreiving and displaying transaction info.
We will want to customize the view, but I haven't yet figured out how to do that.
I hope django-paypal lets you do that in a relatively clean way.

I also implemented an IPN listener, that for now only prints a message to the log
when an IPN request indicating successful payment comes in. Paypal docs warn that
one should guard for duplicates, but it looks like django-paypal already does that.
We should however check that the receiver address, the amount, and item code are correct;
this is yet to be implemented.

Since I can't convince my router to make my computer reachable from the outside internet,
I wasn't yet able to test whether IPN works. I plan to deploy on test.cardstories.org and
check how well that works.