Implement grant_cards_to_player(player, cards) on the service (closes #1003).
This implements a HTTP endpoint on the service which, when called, grants player
a set of cards as passed in in with the arguments.

The endpoint should only be accessible from the local network (django), so that players
can't grant themselves cards simply by POSTing to an url.

Securing the endpoint is left as an excercise to the reader.

Or maybe we'll add a separate task to do that :).