Implement the IPN handler (closes #1007).
This implements the IPN handler such that if the IPN request seems valid,
it grants the player cards from the extra cards pack after he makes the purchase.

It first checks that the request seems "right" - the receiver email,
amount/currency, and status of the payment according to the PayPal guidelines.

PayPal also warns that an IPN request for the same transaction might come in
more than once and you should guard aginst that by storing handled transaction ids.

I didn't bother to do that for two reasons:

- It all looks like django-paypal does that already. It isn't documented, but looking
  at the code that does seem to be the case.

- In our case handling the transaction means granting cards to the player. It really doesn't
  matter whether the handler is invoked more than once per transaction as granting
  player cards that he already "owns" is a noop.

It would also nice to add some logging to the IPN handler - to log when for instance somebody
is trying to cheat by sending fake requests that fail validation, or when a valid request
isn't handled properly because the webservice is having problems handling the request.

I chose to leave that for a separate task for three reasons:

- I don't what the proper way to do logging in django is. I read the documentation but that
  didn't answer my question of what the simplest, idiomatic approach to logging with django is.
  As far as I can tell, we don't do any custom logging from the django app yet, so there is no
  example for me to follow.

- I actually tried using the logging module on our test server instance, but it didn't look like
  what I was sending to the logger was logged in any file, so there's probably some configuration
  that needs to be done and is not just about code.

- All transaction and their statuses will be visible in the PayPal dashboard, so it is not extremely
  important that we also implement logging in our IPN handler as we can always look at what was going
  on in the PayPal dashboard (although with a bit less detail).