Store purchases by users in the django database (closes #1014).
This commit creates a new django model Purchase with three attributes:
ID of the user who made the purchase and a code of the item that was bought,
and a timestamp of the time the purhcase was made.

The purchases table is indexed on user_id and item_code so that it should
be cheap to check whether a particular user has already bought the cards or not.