Hide the 'Buy cards' button for players who already bought cards (closes #1015).
This is sort of a hackish solution that I can't really decide whether I'm happy
with or not. The 'Buy more cards' button is embedded in the game HTML, which is not
part of django, se we can't use a simple {% if... %} block to hide the button
for users who already bought the cards.

The 'proper' solution might be to pass the information about whether the player
bought cards from a django template to the JS client as a javascript variable,
and let the client perform the hiding of the button.
But then again this doesn't feel quite right since the client doesn't really
know anything about buying extra cards, it just so happens that the button was
designed to be part of the game board.

So instead I chose a simpler, if sort of hackish solution:
For users who already bought cards, inject a CSS rule that hides
the button into the header of the django layout.