Implement some logging in the IPN handler (closes #1008).
This sprinkles some logging across the IPN handler using the logging module.
The logs are written to a rotated log file.

Django 1.3 integrates the logging module and has a nice syntax for configuring
loggers in the settings.py file, but since we are using 1.2.5, I had to manually
configure the logger at the bottom of settings.py.