Rename player_id to user_id so that djangoauth doesn't try to authenticate the player (refs #1007).
Djangoauth will try to authenticate any request to "/resource" that contains a param
called 'player_id'. A quick solution was to rename the parameter to 'user_id'.

A better solution would be to move the 'grant_cards_to_player' and any other internal
actions that are only meant to be used for internal ws <-> django communication and
aren't meant to be directly accessible from the public to reside under a separate resource,
perhaps under "/internal" or "/api", where WS plugins wouldn't be run.
I'll create a task for that.