Create an internal WS resource under '/internal' (closes #1018).
The only difference between the "/internal" resource and the regular "/resource" is
that the internal resource doesn't run the plugins; and calls CardstoriesService.handle
method with internal_request parameter set to True. The service knows that if this parameter
is set to True, and only then, it is allowed to handle actions that are marked under ACTIONS_INTERNAL.

The idea is that the "/internal" resource won't be exposed to the outside world (the way "/resource is)
and so we should be able to trust the requests hitting "/internal".

The tests are not something I can be proud of, if anybody has any suggestions, I'd love to improve them.