Add PAYPAL_TEST configuration variable (closes #1019).
The django-paypal module supports a PAYPAL_TEST variable, but it only seems to be
half implemented - the variable does control where the buy button images in the rendered
form are pulled from (www.paypal.com vs www.sandbox.paypal.com), but it doesn't have any
effect on the value of the "action" attribute of the rendered form.

You need to explicitly call form.sandbox if you want the form to point to the sandbox,
even if PAYPAL_TEST is set to True. So I added an if sentence to the template where
the form is rendered so that the proper method gets called, depending on the
value of PAYPAL_TEST.