Add PAYPAL_DEBUG setting variable (closes #1020).
Examining django-paypal's code, it looks like django-paypal uses
the PAYPAL_DEBUG configuration variable to determine which endpoint
(live or sandbox) it sends the verification requests to.

If PAYPAL_DEBUG variable is not set, it defaults to the value of django's
DEBUG variable, which in our case is set to True, causing django-paypal
to send verification requests to the sandbox instead of paypal live.

I didn't notice this earlier because I didn't expect django-paypal configuration
to be such a mess - when switching from sandbox to the real thing,
there are at least three things that we had to do:

- The button form is always rendered to point to the live version,
  regardless of any configuration variables; if you want to point it to the sandbox,
  you need to call .sandbox() rather than .render() on the form.

- The button images are pulled either from live or sandbox version,
  depending on the value of the PAYPAL_TEST variable.

- The PDT postback verification request goes to either live or sandbox,
  depending on the value of the PAYPAL_DEBUG variable (which defaults to DEBUG).

I added PAYPAL_DEBUG to our settings.py and set it to True,
it should be overriden and set to False in local_settings.py on
our production server.

I hope this fixes the 'Error 4011' problem.