Use a shared secret for internal Django -> WS communication (closes #1006).
Requests hitting the webservice's /internal handler are expected
to include a parameter called 'secret' that contains a secret key
which matches the key the webservice is configured with.
If it doesn't match, or the request doesn't contain the 'secret' parameter
at all, the service responds with an unauthorized error.

The service's secret is set up via a command-line parameter
called --internal-secret, while the same secret should be set up on
the django side in local_settings.py as the WEBSERVICE_INTERNAL_SECRET
parameter.