Implement activity notifications opt-out field and django handler (closes #972).
This commit adds a boolean field called 'activity_notifications_disabled' to the
cardstories_userprofile table which defaults to False.

It also adds a handler under /activity-notifications/unsubscribe, which in case the
user is logged in shows him a question "Are you sure you want to unsubscribe?"
with buttons for "Yes" and "No". Upon clicking the "Yes" button, the DB field is
set to True.
If the user isn't logged in, the page instead asks the player to log in and points
him to the /login url with a special return_to=... url parameter set, so that
upon loggin in, the player is returned to the unsubscribe page where he can now
see the "yes/no" question.

I had implement the handling of the return_to parameter, which still seemed easier than
implementing some kind of a unique one-time token scheme that would allow users
to unsubscribe without logging in.

When trying to send out an activity mail using the actual code I got an error
because the mailing.loop.iso_to_datetime function was expecting a datetime string
with fractional seconds but received a datetime string without fractional seconds.

I am not sure why the original code included fractional seconds. My guess is that it
might have been written that way because we were populating the test database (inside the
test_loop.py test in a way that the data included fractional seconds,
while the real webservice DB stores datetimes without fractional seconds.
I changed the tests (and the iso_to_datetime function) to work with
datetime strings without fractional seconds, hoping this will match the production setup.